GpuPointBuffer v1
