+++ pasture-core/src/gpu/mod.rs
@@ -1,2 +1,5 @@
 mod device;
 pub use self::device::*;
+
+mod gpu_point_buffer;
+pub use self::gpu_point_buffer::*;
